requirements.txt
python-dotenv==1.1.0
streamlit==1.45.1
requests
PyMuPdf==1.26.1
langchain==0.3.26
databricks-langchain==0.5.1
mlflow-skinny==3.1.0